hackyeah2024/accounts/serializers.py
from rest_framework import serializers

from hackyeah2024.accounts.models import CustomUser


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = (
            "id",
            "username",
        )
