requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/dev.in
#
amqp==5.2.0
    # via
    #   -r requirements/base.txt
    #   kombu
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
    #   django-stubs
asttokens==2.4.1
    # via
    #   -r requirements/base.txt
    #   stack-data
billiard==4.2.1
    # via
    #   -r requirements/base.txt
    #   celery
celery==5.4.0
    # via -r requirements/base.txt
certifi==2024.8.30
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.7
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
coverage[toml]==7.6.1
    # via pytest-cov
cryptography==43.0.1
    # via
    #   -r requirements/base.txt
    #   social-auth-core
decorator==5.1.1
    # via
    #   -r requirements/base.txt
    #   ipython
defusedxml==0.8.0rc2
    # via
    #   -r requirements/base.txt
    #   python3-openid
    #   social-auth-core
distlib==0.3.8
    # via virtualenv
django==5.1.1
    # via
    #   -r requirements/base.txt
    #   django-dbbackup
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-jazzmin
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-yasg
    #   social-auth-app-django
django-dbbackup==4.2.1
    # via -r requirements/base.txt
django-debug-toolbar==4.4.6
    # via -r requirements/base.txt
django-extensions==3.2.3
    # via -r requirements/base.txt
django-filter==24.3
    # via -r requirements/base.txt
django-jazzmin==3.0.0
    # via -r requirements/base.txt
django-stubs==5.1.0
    # via -r requirements/dev.in
django-stubs-ext==5.1.0
    # via django-stubs
django-templated-mail==1.1.1
    # via
    #   -r requirements/base.txt
    #   djoser
djangorestframework==3.15.2
    # via
    #   -r requirements/base.txt
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.1
    # via
    #   -r requirements/base.txt
    #   djoser
djoser==2.2.3
    # via -r requirements/base.txt
drf-flex-fields==1.0.2
    # via -r requirements/base.txt
drf-yasg==1.21.7
    # via -r requirements/base.txt
environs==11.0.0
    # via -r requirements/base.txt
execnet==2.1.1
    # via pytest-xdist
executing==2.1.0
    # via
    #   -r requirements/base.txt
    #   stack-data
factory-boy==3.3.1
    # via -r requirements/dev.in
faker==30.0.0
    # via factory-boy
filelock==3.16.1
    # via virtualenv
identify==2.6.1
    # via pre-commit
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
inflection==0.5.1
    # via
    #   -r requirements/base.txt
    #   drf-yasg
iniconfig==2.0.0
    # via pytest
ipython==8.27.0
    # via -r requirements/base.txt
jedi==0.19.1
    # via
    #   -r requirements/base.txt
    #   ipython
kombu==5.4.2
    # via
    #   -r requirements/base.txt
    #   celery
marshmallow==3.22.0
    # via
    #   -r requirements/base.txt
    #   environs
matchlib==0.2.3
    # via -r requirements/dev.in
matplotlib-inline==0.1.7
    # via
    #   -r requirements/base.txt
    #   ipython
mypy==1.11.2
    # via -r requirements/dev.in
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
oauthlib==3.2.2
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
packaging==24.1
    # via
    #   -r requirements/base.txt
    #   drf-yasg
    #   marshmallow
    #   pytest
    #   pytest-sugar
parso==0.8.4
    # via
    #   -r requirements/base.txt
    #   jedi
pexpect==4.9.0
    # via
    #   -r requirements/base.txt
    #   ipython
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.8.0
    # via -r requirements/dev.in
prompt-toolkit==3.0.48
    # via
    #   -r requirements/base.txt
    #   click-repl
    #   ipython
psycopg2-binary==2.9.9
    # via -r requirements/base.txt
ptyprocess==0.7.0
    # via
    #   -r requirements/base.txt
    #   pexpect
pure-eval==0.2.3
    # via
    #   -r requirements/base.txt
    #   stack-data
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pygments==2.18.0
    # via
    #   -r requirements/base.txt
    #   ipython
pyjwt==2.9.0
    # via
    #   -r requirements/base.txt
    #   djangorestframework-simplejwt
    #   social-auth-core
pytest==8.3.3
    # via
    #   -r requirements/dev.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-django
    #   pytest-sugar
    #   pytest-xdist
pytest-asyncio==0.24.0
    # via -r requirements/dev.in
pytest-cov==5.0.0
    # via -r requirements/dev.in
pytest-django==4.9.0
    # via -r requirements/dev.in
pytest-sugar==1.0.0
    # via -r requirements/dev.in
pytest-xdist==3.6.1
    # via -r requirements/dev.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   celery
    #   faker
    #   time-machine
python-dotenv==1.0.1
    # via
    #   -r requirements/base.txt
    #   environs
python3-openid==3.2.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
pytz==2024.2
    # via
    #   -r requirements/base.txt
    #   django-dbbackup
    #   drf-yasg
pyyaml==6.0.2
    # via
    #   -r requirements/base.txt
    #   drf-yasg
    #   pre-commit
requests==2.32.3
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
ruff==0.6.8
    # via -r requirements/dev.in
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   asttokens
    #   python-dateutil
social-auth-app-django==5.4.2
    # via
    #   -r requirements/base.txt
    #   djoser
social-auth-core==4.5.4
    # via
    #   -r requirements/base.txt
    #   social-auth-app-django
sqlparse==0.5.1
    # via
    #   -r requirements/base.txt
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via
    #   -r requirements/base.txt
    #   ipython
termcolor==2.4.0
    # via pytest-sugar
time-machine==2.15.0
    # via -r requirements/dev.in
traitlets==5.14.3
    # via
    #   -r requirements/base.txt
    #   ipython
    #   matplotlib-inline
types-pyyaml==6.0.12.20240917
    # via django-stubs
typing-extensions==4.12.2
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
tzdata==2024.2
    # via
    #   -r requirements/base.txt
    #   celery
    #   kombu
uritemplate==4.1.1
    # via
    #   -r requirements/base.txt
    #   drf-yasg
urllib3==2.2.3
    # via
    #   -r requirements/base.txt
    #   requests
vine==5.1.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
virtualenv==20.26.5
    # via pre-commit
wcwidth==0.2.13
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit
